03_mcp-adk-a2a/src/session1/requirements.txt
fastmcp==0.1.0
requests==2.31.0
python-dotenv==1.0.0